Properly painted and textured all my models
Except for the rug, cause fuck that one. I made too many patterns and it was a pain to paint, so it might look more shitty.
Reimported the models and fixed the UVs cuz I updated SubstancePainter to 2.2.3.
All the textured & painted models can be found in the ModelPrefab Folder, ready to be dropped into game.
PainTextures folder is for SubstancePainter files.
StockImage folder is for downloaded images to be used as materials.